Move classes de teste de crud
