+++ src/br/com/caelum/financas/modelo/crud/TestaEdicaoConta.java
@@ -1,4 +1,4 @@
-package br.com.caelum.financas.modelo.teste;
+package br.com.caelum.financas.modelo.crud;
 
 import javax.persistence.EntityManager;
 
@@ -6,16 +6,20 @@ import br.com.caelum.financas.dao.ContaDao;
 import br.com.caelum.financas.infra.JPAUtil;
 import br.com.caelum.financas.modelo.Conta;
 
-public class TestaInsereConta {
+public class TestaEdicaoConta {
 
 	public static void main(String[] args) {
-		Conta conta = new Conta("Alexandre Gama", "1234567", "123");
+		Conta novaConta = new Conta("Alexandre Gama", "1234567", "123");
 		
 		EntityManager manager = new JPAUtil().getEntityManager();
+		ContaDao dao = new ContaDao(manager);
 		manager.getTransaction().begin();
+		dao.adiciona(novaConta);
 		
-		ContaDao contaDao = new ContaDao(manager);
-		contaDao.adiciona(conta);
+		Conta conta = dao.buscaPorId(novaConta.getId());
+		conta.setTitular("Gama");
+		manager.merge(conta);
+		
 		manager.getTransaction().commit();
 		manager.close();
 	}
