+++ src/br/com/caelum/financas/modelo/crud/TestaInsereConta.java
@@ -1,4 +1,4 @@
-package br.com.caelum.financas.modelo.teste;
+package br.com.caelum.financas.modelo.crud;
 
 import javax.persistence.EntityManager;
 
@@ -6,20 +6,16 @@ import br.com.caelum.financas.dao.ContaDao;
 import br.com.caelum.financas.infra.JPAUtil;
 import br.com.caelum.financas.modelo.Conta;
 
-public class TestaEdicaoConta {
+public class TestaInsereConta {
 
 	public static void main(String[] args) {
-		Conta novaConta = new Conta("Alexandre Gama", "1234567", "123");
+		Conta conta = new Conta("Alexandre Gama", "1234567", "123");
 		
 		EntityManager manager = new JPAUtil().getEntityManager();
-		ContaDao dao = new ContaDao(manager);
 		manager.getTransaction().begin();
-		dao.adiciona(novaConta);
 		
-		Conta conta = dao.buscaPorId(novaConta.getId());
-		conta.setTitular("Gama");
-		manager.merge(conta);
-		
+		ContaDao contaDao = new ContaDao(manager);
+		contaDao.adiciona(conta);
 		manager.getTransaction().commit();
 		manager.close();
 	}
